# Add transient optimization problems 