@kliegeois 2024-09-05
The last commit temporary disable the Tempus driven test as the check gradient verifications are not OK while using Tempus for now: a singular W is generated and associated solves fail. Using SDBC instead of DBC fix that but then none of the gradients are correct even the ROL driven ones.

I am working on this and will create a different PR. This PR then only add ROL driven transient optimization without SDBC with scalar and distributed parameters, and integrated over time or final time step response.